+++ src/collections/vscode/index.ts
@@ -3,6 +3,13 @@ export { CollectionTreeDataProvider } from './collection-tree-provider';
 export { CollectionNavigationService } from './navigation-service';
 export { registerCollections } from './register-collections';
 export type { CollectionsRegistration } from './register-collections';
+export { VsCodeCollectionFilesystem } from './mutation-filesystem';
+export { CollectionTreeDragAndDropController } from './collection-dnd-controller';
+export {
+  buildNewRequestDestinations,
+  openNewRequestDialog,
+} from './new-request-dialog';
+export type { NewRequestDialogResult } from './new-request-dialog';
 export {
   NodeApiFileReader,
   VsCodeApiFileReader,